Update README.md
Abstract: The petroleum industry plays a crucial role in global energy supply, making efficient and secure management of its supply chain essential. Blockchain technology has emerged as a promising solution for enhancing transparency, traceability, and security in various industries. This project aims to explore the application of  blockchain technology in monitoring the movement of petroleum from extraction to the end consumer, thereby revolutionizing the petroleum supply chain. The findings and recommendations can serve as a foundation for real-world implementations and pave the way for a more accountable petroleum industry.